cleaner handling of grobid service shutdown

+++ start_queues.sh
@@ -5,7 +5,8 @@ for pid in $RUNNING; do
     echo "Killing queue process [PID $pid]"
     kill -9 $pid
 done
-kill $(ps ax | grep grobid-service | grep -v grep | awk '{print $1}')
+
+if [[ $(ps ax | grep grobid-service | grep -v grep) ]]; then kill $(ps ax | grep grobid-service | grep -v grep | awk '{print $1}'); fi
 
 php public/index.php queue doctrine bibtex --start &
 echo "Started Bibtex queue [PID $!]"
